deleted duplicate errorModal
